+++ packages/server/lib/src/server/query.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 
+import 'package:engine_io_shared/exceptions.dart';
 import 'package:engine_io_shared/transports.dart';
 import 'package:universal_io/io.dart' hide Socket;
 
-import 'package:engine_io_server/src/server/exception.dart';
 import 'package:engine_io_server/src/server/server.dart';
 
 /// Contains the parameters extracted from HTTP queries.