+++ packages/server/lib/src/server/server.dart
@@ -1,13 +1,12 @@
 import 'dart:async';
 
+import 'package:engine_io_shared/exceptions.dart';
 import 'package:engine_io_shared/packets.dart';
 import 'package:engine_io_shared/transports.dart';
 import 'package:universal_io/io.dart' hide Socket;
 
-import 'package:engine_io_server/src/exception.dart';
 import 'package:engine_io_server/src/server/client_manager.dart';
 import 'package:engine_io_server/src/server/configuration.dart';
-import 'package:engine_io_server/src/server/exception.dart';
 import 'package:engine_io_server/src/server/query.dart';
 import 'package:engine_io_server/src/server/socket.dart';
 import 'package:engine_io_server/src/transports/polling/polling.dart';