+++ packages/server/lib/src/server/upgrade.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 
+import 'package:engine_io_shared/exceptions.dart';
 import 'package:engine_io_shared/transports.dart';
 
 import 'package:engine_io_server/src/server/socket.dart';
-import 'package:engine_io_server/src/transports/exception.dart';
 import 'package:engine_io_server/src/transports/transport.dart';
 
 /// Represents the status of a transport upgrade.