+++ packages/server/lib/src/transports/polling/polling.dart
@@ -1,12 +1,11 @@
 import 'dart:collection';
 import 'dart:convert';
 
+import 'package:engine_io_shared/exceptions.dart';
 import 'package:engine_io_shared/packets.dart';
 import 'package:engine_io_shared/transports.dart';
 import 'package:universal_io/io.dart' hide Socket;
 
-import 'package:engine_io_server/src/transports/exception.dart';
-import 'package:engine_io_server/src/transports/polling/exception.dart';
 import 'package:engine_io_server/src/transports/transport.dart';
 import 'package:engine_io_server/src/transports/websocket/websocket.dart';