+++ packages/server/lib/src/transports/transport.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:engine_io_shared/exceptions.dart';
 import 'package:engine_io_shared/packets.dart';
 import 'package:engine_io_shared/transports.dart';
 import 'package:universal_io/io.dart' hide Socket;
@@ -7,7 +8,6 @@ import 'package:universal_io/io.dart' hide Socket;
 import 'package:engine_io_server/src/server/configuration.dart';
 import 'package:engine_io_server/src/server/socket.dart';
 import 'package:engine_io_server/src/server/upgrade.dart';
-import 'package:engine_io_server/src/transports/exception.dart';
 import 'package:engine_io_server/src/transports/heartbeat_manager.dart';
 import 'package:engine_io_server/src/transports/polling/polling.dart';