+++ packages/server/test/server/upgrade_test.dart
@@ -1,17 +1,16 @@
 import 'dart:async';
 
 import 'package:async/async.dart';
+import 'package:engine_io_shared/exceptions.dart';
 import 'package:engine_io_shared/packets.dart';
 import 'package:engine_io_shared/transports.dart';
 import 'package:test/test.dart';
 import 'package:universal_io/io.dart' hide Socket, SocketException;
 
 import 'package:engine_io_server/src/server/configuration.dart';
-import 'package:engine_io_server/src/server/exception.dart';
 import 'package:engine_io_server/src/server/server.dart';
 import 'package:engine_io_server/src/server/socket.dart';
 import 'package:engine_io_server/src/server/upgrade.dart';
-import 'package:engine_io_server/src/transports/exception.dart';
 import 'package:engine_io_server/src/transports/websocket/websocket.dart';
 import '../matchers.dart';
 import '../shared.dart';