+++ packages/shared/lib/src/transports/websocket/exception.dart
@@ -1,11 +1,9 @@
-import 'package:universal_io/io.dart';
-
-import 'package:engine_io_server/src/transports/exception.dart';
+import 'package:engine_io_shared/src/transports/exception.dart';
 
 /// An exception that occurred on a websocket transport.
 class WebSocketTransportException extends TransportException {
   @override
-  bool get isSuccess => statusCode == WebSocketStatus.normalClosure;
+  bool get isSuccess => statusCode == 1000; // Normal closure
 
   /// Creates an instance of `WebSocketTransportException`.
   const WebSocketTransportException({
@@ -15,13 +13,13 @@ class WebSocketTransportException extends TransportException {
 
   /// The server failed to decode a packet.
   static const decodingPacketFailed = WebSocketTransportException(
-    statusCode: WebSocketStatus.policyViolation,
+    statusCode: 1008,
     reasonPhrase: 'Failed to decode packet.',
   );
 
   /// The client sent data of an unknown data type.
   static const unknownDataType = WebSocketTransportException(
-    statusCode: WebSocketStatus.policyViolation,
+    statusCode: 1008,
     reasonPhrase:
         'Could not recognise the kind of data that has been sent over.',
   );