+++ packages/shared/CHANGELOG.md
@@ -13,3 +13,8 @@
     - `noop`: `NoopPacket`
 - Transports:
   - Define the available connection types in the `ConnectionType` enum.
+- Define the `EngineException` class:
+  - Define the `SocketException` class.
+  - Define the `TransportException` class:
+    - Define the `PollingTransportException` class.
+    - Define the `WebSocketTransportException` class.